feat: Add content to grilled-teriyaki-shrimp.html
This commit adds the entire HTML content to the Grilled Teriyaki Shrimp
and Pineapple Skewers recipe page. starting with an image of the dish,
and below , adding a description of the dish, an <ul> for ingredients
and an <ol> for cooking steps.

Minor fix: indentation was fixed to 2 spaces through all html files